# D_labs/lab_d5/lab_d5_openapi.json
{"openapi":"3.1.0","info":{"title":"Flights API Service","version":"0.1.0"},"paths":{"/cities":{"get":{"summary":"Get Cities","operationId":"get_cities_cities_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/airports":{"get":{"summary":"Get Airports","operationId":"get_airports_airports_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/cities/{city_name}/airports":{"get":{"summary":"Get Airports In City","operationId":"get_airports_in_city_cities__city_name__airports_get","parameters":[{"name":"city_name","in":"path","required":true,"schema":{"type":"string","title":"City Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/airports/{code}/schedule/inbound":{"get":{"summary":"Get Inbound Schedule","operationId":"get_inbound_schedule_airports__code__schedule_inbound_get","parameters":[{"name":"code","in":"path","required":true,"schema":{"type":"string","title":"Code"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/airports/{code}/schedule/outbound":{"get":{"summary":"Get Outbound Schedule","operationId":"get_outbound_schedule_airports__code__schedule_outbound_get","parameters":[{"name":"code","in":"path","required":true,"schema":{"type":"string","title":"Code"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/check-in":{"post":{"summary":"Check In","operationId":"check_in_check_in_post","parameters":[{"name":"ticket_no","in":"query","required":true,"schema":{"type":"string","title":"Ticket No"}},{"name":"flight_id","in":"query","required":true,"schema":{"type":"integer","title":"Flight Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/search":{"post":{"summary":"Search Flights","operationId":"search_flights_search_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/bookings":{"post":{"summary":"Create Booking","operationId":"create_booking_bookings_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookingRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"BookingRequest":{"properties":{"passenger_id":{"type":"string","title":"Passenger Id"},"passenger_name":{"type":"string","title":"Passenger Name"},"booking_class":{"type":"string","title":"Booking Class"},"flight_ids":{"items":{"type":"integer"},"type":"array","title":"Flight Ids"}},"type":"object","required":["passenger_id","passenger_name","booking_class","flight_ids"],"title":"BookingRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"SearchRequest":{"properties":{"from_point":{"type":"string","title":"From Point"},"to_point":{"type":"string","title":"To Point"},"departure_date":{"type":"string","title":"Departure Date"},"booking_class":{"type":"string","title":"Booking Class"},"max_connections":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Max Connections"},"limit":{"type":"integer","title":"Limit","default":50}},"type":"object","required":["from_point","to_point","departure_date","booking_class"],"title":"SearchRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}